+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Character/KCCStateMachine.cs
@@ -20,10 +20,14 @@ using System;
 using nickmaltbie.OpenKCC.Character.Action;
 using nickmaltbie.OpenKCC.Character.Config;
 using nickmaltbie.OpenKCC.Character.Events;
-using nickmaltbie.OpenKCC.FSM;
-using nickmaltbie.OpenKCC.FSM.Attributes;
+using nickmaltbie.OpenKCC.Environment.MovingGround;
 using nickmaltbie.OpenKCC.Utils;
+using nickmaltbie.StateMachineUnity;
+using nickmaltbie.StateMachineUnity.Attributes;
+using nickmaltbie.StateMachineUnity.Event;
+using nickmaltbie.StateMachineUnity.Fixed;
 using UnityEngine;
+using UnityEngine.Animations;
 using UnityEngine.InputSystem;
 using static nickmaltbie.OpenKCC.Utils.KCCUtils;
 
@@ -33,12 +37,15 @@ namespace nickmaltbie.OpenKCC.Character
     /// Have a character controller push any dynamic rigidbody it hits
     /// </summary>
     [RequireComponent(typeof(Rigidbody))]
-    public class KCCStateMachine : FixedStateMachineBehaviour, IKCCConfig, IJumping
+    public class KCCStateMachine : FixedSMAnim, IKCCConfig, IJumping
     {
-        /// <summary>
-        /// Unity service for managing unit inputs in a testable manner.
-        /// </summary>
-        public IUnityService unityService = UnityService.Instance;
+        public const string IdleAnimState = "Idle";
+        public const string JumpAnimState = "Jump";
+        public const string LandingAnimState = "Landing";
+        public const string WalkingAnimState = "Walking";
+        public const string SlidingAnimState = "Sliding";
+        public const string FallingAnimState = "Falling";
+        public const string LongFallingAnimState = "Long Falling";
 
         [Header("Input Controls")]
 
@@ -240,46 +247,87 @@ namespace nickmaltbie.OpenKCC.Character
         /// </summary>
         public Vector3 InputMovement { get; private set; }
 
+        /// <summary>
+        /// Parent constraints to attach position to floor.
+        /// </summary>
+        private ParentConstraint parentConstraint;
+
+        /// <summary>
+        /// Floor Constraint used for parent constraint to move with ground/
+        /// </summary>
+        private ConstraintSource floorConstraint;
+
         [InitialState]
+        [Animation(IdleAnimState, 0.35f, true)]
         [Transition(typeof(MoveInput), typeof(WalkingState))]
         [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
         [Transition(typeof(LeaveGroundEvent), typeof(FallingState))]
-        [MovementSettings(AllowVelocity = false, AllowWalk = false)]
-        [OnFixedUpdate(nameof(SnapPlayerDown))]
+        [Transition(typeof(JumpEvent), typeof(JumpState))]
+        [MovementSettings(AllowVelocity = false, AllowWalk = false, SnapPlayerDown = true)]
         public class IdleState : State { }
 
+        [ApplyGravity]
+        [Animation(JumpAnimState, 0.1f, true)]
+        [TransitionOnAnimationComplete(typeof(FallingState), 0.15f, true)]
+        [AnimationTransition(typeof(GroundedEvent), typeof(LandingState), 0.35f, true, 0.25f)]
+        [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
+        [MovementSettings(AllowVelocity = true, AllowWalk = true)]
+        public class JumpState : State { }
+
+        [Animation(LandingAnimState, 0.1f, true)]
+        [TransitionOnAnimationComplete(typeof(IdleState), 0.25f, true)]
+        [AnimationTransition(typeof(MoveInput), typeof(WalkingState), 0.35f, true)]
+        [AnimationTransition(typeof(JumpEvent), typeof(JumpState), 0.35f, true)]
+        [Transition(typeof(LeaveGroundEvent), typeof(FallingState))]
+        [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
+        [MovementSettings(AllowVelocity = false, AllowWalk = true, SnapPlayerDown = true)]
+        public class LandingState : State { }
+
+        [Animation(WalkingAnimState, 0.1f, true)]
+        [Transition(typeof(JumpEvent), typeof(JumpState))]
         [Transition(typeof(StopMoveInput), typeof(IdleState))]
         [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
         [Transition(typeof(LeaveGroundEvent), typeof(FallingState))]
-        [MovementSettings(AllowVelocity = false, AllowWalk = true)]
-        [OnFixedUpdate(nameof(SnapPlayerDown))]
+        [MovementSettings(AllowVelocity = false, AllowWalk = true, SnapPlayerDown = true)]
         public class WalkingState : State { }
 
-        [Transition(typeof(LeaveGroundEvent), typeof(FallingState))]
-        [Transition(typeof(GroundedEvent), typeof(IdleState))]
         [ApplyGravity]
-        [MovementSettings(AllowVelocity = true, AllowWalk = true)]
-        [OnFixedUpdate(nameof(SnapPlayerDown))]
+        [Animation(SlidingAnimState, 0.35f, true)]
+        [Transition(typeof(JumpEvent), typeof(JumpState))]
+        [Transition(typeof(LeaveGroundEvent), typeof(FallingState))]
+        [AnimationTransition(typeof(GroundedEvent), typeof(LandingState), 0.35f, true, 0.25f)]
+        [MovementSettings(AllowVelocity = true, AllowWalk = true, SnapPlayerDown = true)]
         public class SlidingState : State { }
 
-        [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
-        [Transition(typeof(GroundedEvent), typeof(IdleState))]
         [ApplyGravity]
+        [Animation(FallingAnimState, 0.1f, true)]
+        [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
+        [AnimationTransition(typeof(GroundedEvent), typeof(LandingState), 0.35f, true, 0.25f)]
+        [TransitionAfterTime(typeof(LongFallingState), 2.0f)]
         [MovementSettings(AllowVelocity = true, AllowWalk = true)]
         public class FallingState : State { }
+
+        [ApplyGravity]
+        [Animation(LongFallingAnimState, 0.1f, true)]
+        [Transition(typeof(JumpEvent), typeof(JumpState))]
+        [Transition(typeof(SteepSlopeEvent), typeof(SlidingState))]
+        [AnimationTransition(typeof(GroundedEvent), typeof(LandingState), 0.35f, true, 1.0f)]
+        [MovementSettings(AllowVelocity = true, AllowWalk = true)]
+        public class LongFallingState : State { }
 
         /// <inheritdoc/>
         public override void FixedUpdate()
         {
+            while (parentConstraint.sourceCount > 0)
+            {
+                parentConstraint.RemoveSource(0);
+            }
+
             // Push player out of overlapping objects
             PushOutOverlapping();
 
             // Update grounded state
-            groundedState.CheckGrounded(this, transform.position, transform.rotation);
-            RaiseEvent(
-                !groundedState.StandingOnGround ? LeaveGroundEvent.Instance :
-                groundedState.Sliding ? SteepSlopeEvent.Instance :
-                    GroundedEvent.Instance as IEvent);
+            UpdateGroundedState();
 
             if (CurrentState == typeof(FallingState))
             {
@@ -300,6 +348,83 @@ namespace nickmaltbie.OpenKCC.Character
             ApplyMovement();
 
             base.FixedUpdate();
+
+            UpdateMovingGround();
+        }
+
+        /// <summary>
+        /// Update the grounded state of the kinematic character controller.
+        /// </summary>
+        public void UpdateGroundedState()
+        {
+            groundedState.CheckGrounded(this, transform.position, transform.rotation);
+            IEvent groundedEvent;
+            if (!groundedState.StandingOnGround)
+            {
+                groundedEvent = LeaveGroundEvent.Instance;
+            }
+            else if (groundedState.Sliding)
+            {
+                groundedEvent = SteepSlopeEvent.Instance;
+            }
+            else
+            {
+                groundedEvent = GroundedEvent.Instance;
+            }
+
+            RaiseEvent(groundedEvent);
+        }
+
+        /// <summary>
+        /// Update the moving ground state of the KCC State machine.
+        /// </summary>
+        public void UpdateMovingGround()
+        {
+            groundedState.CheckGrounded(this, transform.position, transform.rotation);
+            bool movingGround = groundedState.StandingOnGround &&
+                groundedState.Floor?.GetComponent<IMovingGround>() != null;
+            parentConstraint.constraintActive = movingGround;
+            parentConstraint.translationAtRest = transform.position;
+            parentConstraint.rotationAtRest = transform.rotation.eulerAngles;
+
+            if (movingGround)
+            {
+                Transform floorTransform = groundedState.Floor.transform;
+                floorConstraint.sourceTransform = floorTransform;
+                floorConstraint.weight = 1.0f;
+                parentConstraint.AddSource(floorConstraint);
+
+                Vector3 relativePos = transform.position - floorTransform.position;
+                Vector3 localRelativePos = floorTransform.InverseTransformDirection(relativePos);
+                parentConstraint.SetTranslationOffset(0, localRelativePos);
+            }
+            else
+            {
+                floorConstraint = default;
+            }
+        }
+
+        /// <summary>
+        /// Gets the velocity of the ground the player is standing on where the player is currently
+        /// </summary>
+        /// <returns>The velocity of the ground at the point the player is standing on</returns>
+        public Vector3 GetGroundVelocity(GameObject floor, Vector3 groundHitPosition)
+        {
+            Vector3 groundVelocity = Vector3.zero;
+            IMovingGround movingGround = floor?.GetComponent<IMovingGround>();
+            if (movingGround != null && !movingGround.AvoidTransferMomentum())
+            {
+                // Weight movement of ground by ground movement weight
+                float velocityWeight =
+                    movingGround.GetMovementWeight(groundHitPosition, Velocity);
+                float transferWeight =
+                    movingGround.GetTransferMomentumWeight(groundHitPosition, Velocity);
+                groundVelocity = movingGround.GetVelocityAtPoint(groundHitPosition);
+                groundVelocity *= velocityWeight;
+                groundVelocity *= transferWeight;
+            }
+
+            return groundVelocity;
         }
 
         /// <summary>
@@ -325,6 +450,12 @@ namespace nickmaltbie.OpenKCC.Character
             {
                 Velocity = Vector3.zero;
             }
+
+            // Snap player down if requested
+            if (moveSettings?.SnapPlayerDown ?? false)
+            {
+                SnapPlayerDown();
+            }
         }
 
         /// <summary>
@@ -343,8 +474,15 @@ namespace nickmaltbie.OpenKCC.Character
         /// <summary>
         /// Configure kcc state machine operations.
         /// </summary>
-        public void Awake()
+        public override void Awake()
         {
+            base.Awake();
+
+            parentConstraint = gameObject.AddComponent<ParentConstraint>();
+            parentConstraint.constraintActive = false;
+            parentConstraint.translationAxis = Axis.X | Axis.Y | Axis.Z;
+            parentConstraint.rotationAxis = Axis.None;
+
             GetComponent<Rigidbody>().isKinematic = true;
             jumpAction.Setup(groundedState, this, this);
 
@@ -358,11 +496,18 @@ namespace nickmaltbie.OpenKCC.Character
         /// <inheritdoc/>
         public override void Update()
         {
-            Vector2 moveVector = PlayerInputUtils.playerMovementState == PlayerInputState.Deny ?
-                moveVector = Vector3.zero : moveAction.action.ReadValue<Vector2>();
+            bool denyMovement = PlayerInputUtils.playerMovementState == PlayerInputState.Deny;
+            Vector2 moveVector = denyMovement ? Vector3.zero : moveAction.action.ReadValue<Vector2>();
             InputMovement = new Vector3(moveVector.x, 0, moveVector.y);
-            RaiseEvent(InputMovement.magnitude >= KCCUtils.Epsilon ?
-                MoveInput.Instance as IEvent : StopMoveInput.Instance as IEvent);
+            bool moving = InputMovement.magnitude >= KCCUtils.Epsilon;
+            RaiseEvent(moving ? MoveInput.Instance : StopMoveInput.Instance);
+
+            float moveX = AttachedAnimator.GetFloat("MoveX");
+            float moveY = AttachedAnimator.GetFloat("MoveY");
+            moveX = Mathf.Lerp(moveX, moveVector.x, 4 * unityService.deltaTime);
+            moveY = Mathf.Lerp(moveY, moveVector.y, 4 * unityService.deltaTime);
+            AttachedAnimator.SetFloat("MoveX", moveX);
+            AttachedAnimator.SetFloat("MoveY", moveY);
 
             jumpAction.Update();
             base.Update();
@@ -440,8 +585,8 @@ namespace nickmaltbie.OpenKCC.Character
         /// <inheritdoc/>
         public void ApplyJump(Vector3 velocity)
         {
-            Velocity = velocity;
-            RaiseEvent(LeaveGroundEvent.Instance);
+            Velocity = velocity + GetGroundVelocity(groundedState.Floor, groundedState.GroundHitPosition);
+            RaiseEvent(JumpEvent.Instance);
         }
 
         /// <summary>
@@ -457,12 +602,17 @@ namespace nickmaltbie.OpenKCC.Character
             /// <summary>
             /// Allow movement by normal velocity.
             /// </summary>
-            public bool AllowVelocity = true;
+            public bool AllowVelocity = false;
 
             /// <summary>
             /// Allow movement by player input movement.
             /// </summary>
-            public bool AllowWalk = true;
+            public bool AllowWalk = false;
+
+            /// <summary>
+            /// Should the player be snapped down after moving.
+            /// </summary>
+            public bool SnapPlayerDown = false;
         }
     }
 }