+++ Packages/com.nickmaltbie.openkcc/Tests/EditMode/Environment/FixedRigidbodySetTests.cs
@@ -18,8 +18,8 @@
 
 using Moq;
 using nickmaltbie.OpenKCC.Environment;
-using nickmaltbie.OpenKCC.TestCommon;
 using nickmaltbie.OpenKCC.Utils;
+using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
 using UnityEngine;