+++ Packages/com.nickmaltbie.openkcc/Tests/EditMode/Utils/CapsuleMakerTests.cs
@@ -16,7 +16,7 @@
 // ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using nickmaltbie.OpenKCC.TestCommon;
+using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
 using static nickmaltbie.OpenKCC.Utils.CapsuleMaker;