+++ Packages/com.nickmaltbie.openkcc/Tests/EditMode/Utils/PhysicsWrapperTests.cs
@@ -19,8 +19,8 @@
 using System;
 using System.Linq;
 using System.Reflection;
-using nickmaltbie.OpenKCC.TestCommon;
 using nickmaltbie.OpenKCC.Utils;
+using nickmaltbie.TestUtilsUnity.Tests.TestCommon;
 using NUnit.Framework;
 using UnityEngine;